requirements.txt
opencv-python
face-recognition                                                                
Pillow
pymongo
python-dotenv
customtkinter
tkinter
pydantic